OrderedConcurrentBuffer: reintroduce global sync during flushing
it causes some performance loss, but simplifies the code a lot.
Performance loss explanation:
if a non-head bucket is closed while head bucket thread performs
flushing sequence, the non-head bucket thread will have to wait for
buffer's monitor unnecessarily: once it acquires the buffer's monitor,
it will notice that the buckets are already flushed (line 132) and
return from the method.